Move PrivateComponent, PublicComponent util functions to react-utils
